requirements.txt
quart
quart-corspip
requests
beautifulsoup4
pdfminer.six
tiktoken